Cleaning up code + legion shit
