+++ src/parser/warrior/fury/CombatLogParser.js
@@ -1,43 +1,38 @@
 import CoreCombatLogParser from 'parser/core/CombatLogParser';
 import DamageDone from 'parser/shared/modules/DamageDone';
-
+// Core
+import Enrage from './modules/core/Enrage';
+import Rampage from './modules/core/Rampage';
+// Features
 import Checklist from './modules/features/checklist/Module';
 import Abilities from './modules/Abilities';
 import AlwaysBeCasting from './modules/features/AlwaysBeCasting';
 import CooldownThroughputTracker from './modules/features/CooldownThroughputTracker';
 import SpellUsable from './modules/features/SpellUsable';
-
-import EnrageUptime from './modules/buffdebuff/EnrageUptime';
-import FrothingBerserkerUptime from './modules/buffdebuff/FrothingBerserkerUptime';
-import Juggernaut from './modules/buffdebuff/Juggernaut';
-
-import MissedRampage from './modules/spells/MissedRampage';
-import RampageFrothingBerserker from './modules/features/RampageFrothingBerserker';
-import RampageCancelled from './modules/features/RampageCancelled';
+// Talents
+import FrothingBerserker from './modules/talents/FrothingBerserker';
 import AngerManagement from './modules/talents/AngerManagement';
-import FuriousSlashTimesByStacks from './modules/talents/FuriousSlashTimesByStacks';
-import FuriousSlashUptime from './modules/talents/FuriousSlashUptime';
+import FuriousSlash from './modules/talents/FuriousSlash';
 
 class CombatLogParser extends CoreCombatLogParser {
   static specModules = {
-    damageDone: [DamageDone, {showStatistic: true}],
+    damageDone: [DamageDone, { showStatistic: true }],
 
+    // Core
+    enrage: Enrage,
+    rampage: Rampage,
+
+    // Features
     checklist: Checklist,
     abilities: Abilities,
     alwaysBeCasting: AlwaysBeCasting,
     cooldownThroughputTracker: CooldownThroughputTracker,
     spellUsable: SpellUsable,
 
-    enrageUptime: EnrageUptime,
-    frothingBerserkerUptime: FrothingBerserkerUptime,
-    juggernaut: Juggernaut,
-
-    missedRampage: MissedRampage,
-    rampageFrothingBerserker: RampageFrothingBerserker,
-    rampageCancelled: RampageCancelled,
+    // Talents
+    frothingBerserker: FrothingBerserker,
     angerManagement: AngerManagement,
-    furiousSlashTimesByStacks: FuriousSlashTimesByStacks,
-    furiousSlashUptime: FuriousSlashUptime,
+    furiousSlash: FuriousSlash,
 
     // Overrides default PrePotion
     //prePotion: PrePotion, TODO: Update this to BFA and ensure it works properly with parser/core/Modules/Features/Checklist2/PreparationRuleAnalyzer.js
