+++ src/parser/warrior/fury/modules/core/Enrage.js
@@ -1,0 +1,49 @@
+import React from 'react';
+
+import Analyzer from 'parser/core/Analyzer';
+
+import SPELLS from 'common/SPELLS/index';
+import SpellIcon from 'common/SpellIcon';
+import SpellLink from 'common/SpellLink';
+import { formatPercentage } from 'common/format';
+import StatisticBox from 'interface/others/StatisticBox';
+
+class Enrage extends Analyzer {
+
+  get uptime() {
+    return this.selectedCombatant.getBuffUptime(SPELLS.ENRAGE.id) / this.owner.fightDuration;
+  }
+
+  get suggestionThresholds() {
+    return {
+      actual: this.uptime,
+      isLessThan: {
+        minor: 0.7,
+        average: 0.65,
+        major: 0.6,
+      },
+      style: 'percentage',
+    };
+  }
+
+  suggestions(when) {
+    when(this.suggestionThresholds).addSuggestion((suggest, actual, recommended) => {
+      return suggest(<>Your <SpellLink id={SPELLS.ENRAGE.id} /> uptime can be improved.</>)
+        .icon(SPELLS.ENRAGE.icon)
+        .actual(`${formatPercentage(actual)}% Enrage uptime`)
+        .recommended(`>${formatPercentage(recommended)}% is recommended`);
+    });
+  }
+
+  statistic() {
+    return (
+      <StatisticBox
+        icon={<SpellIcon id={SPELLS.ENRAGE.id} />}
+        value={`${formatPercentage(this.uptime)} %`}
+        label="Enrage uptime"
+      />
+    );
+  }
+}
+
+export default Enrage;
